Pin numpy for CI
With the recent release of numpy 1.20.0 a large number of deprecation
warnings are being raised about the deprecation of type aliases for
stdlib numerical types. This seems to be causing crashes in the nbsphinx
plugin for the tutorials job. To avoid this issue until qiskit has a
release that doesn't raise warnings this commit pins numpy used in CI to
the latest in the 1.19.x series.